ai-video-factory/requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
httpx>=0.28.0
pydantic>=2.0.0
Pillow>=10.0.0
yt-dlp>=2025.0.0

# Optional heavy deps used by visual_service utilities:
# torch
# diffusers
# transformers
# accelerate
